requirements.txt
httpx>=0.27
feedparser>=6.0
pyyaml>=6.0
apscheduler>=3.10,<4
beautifulsoup4>=4.12
starlette>=0.36
uvicorn>=0.27
